Accurate synapse counts in dry-run (respect src target) (#169)
## Context
This PR addresses the discrepancy we were seeing in synapse counts in
Dry-Run mode, as reported in

https://bbpteam.epfl.ch/project/spaces/pages/viewpage.action?spaceKey=BBPHPC&title=Discrepancies+in+synapse+count+for+dry+run+estimate

The difference was due to the simplified counting scheme which
disregarded the "connection_override"s source.

## Scope

This PR fixes the situation by couting synapses in a way that closely
resembles a true run, in the sense that we:

- Get all the pairs (sgid, counts) for a given set of tgids
- for each group of connections per tgid
   - collect sgids and intersect with the source target
   - filter connections by the gids which passed the intersection
   - mark the connection as "seen"
   - add the synapse count in the connection

**Change in synapse grouping**

Besides the immediate fix, this PR significantly changes how synapses
are grouped. Instead if differentiating between EXC and INH synapses,
which had similar memory footprints, one differentiates between
connection types, i.e. Synaptic, GapJunctions, Glia, etc...

Using this grouping one can still be very accurate in terms of memory
usage while much simplifying and speeding up the counting.

## Testing
Using `sscx-v7-plasticity` as the reference, where we would see an
impressive discrepancy, we now have:
```
[INFO]  - Estimated synapse memory usage (MB):
[VERB]  -> +========================== Synapse Count ===========================+
[VERB]  -> |               Synapse Type               |   Count    |   Memory   |
[VERB]  -> +--------------------------------------------------------------------+
[VERB]  -> | ConnectionTypes.Synaptic                 |       5526 |   10.25 MB |
[VERB]  -> +--------------------------------------------------------------------+
```
Notice how close the estimate (5526) is to reality (5581)

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or
files) are good
* [x] Unit tests added

---------

Co-authored-by: Antonio Bellotta <antonio.bellotta@epfl.ch>