Fix multiple issues with Memory Load Balance for MMB run (#185)
## Context
As detailed in
[https://bbpteam.epfl.ch/project/issues/browse/BBPBGLIB-1152](https://bbpteam.epfl.ch/project/issues/browse/BBPBGLIB-1152),
multiple issues came up when trying to use the `neurodamus` memory load
balance to run the MMB circuit.

In particular, the issues were a `list index out of range` error (that
was caused by ranks/cycle sometimes not having any gids assigned) and an
`Unpickling` error that is still misterious in its exact mechanics but
that was seemingly caused by a corrupted broadcast of the allocation
data.

## Scope
This MR implements a huge reworking of the memory load balance workflow:
- the `distribute_cells` function is now a subroutine of
`distribute_cells_with_validation`. This new function implements
directly a check to make sure all ranks/cycle have at least one gid
assigned. Furthermore, it
integrates a dynamic batch size calculation in order to speedup the
distribution.
- the `import_allocation_file` has been completely reworked to integrate
both a caching mechanism (to avoid reloading the allocation file in
every cycle) but also now return only the allocation table for the
appropriate rank.
- the load balance memory workflow now correctly supports
multi-population circuit
- disable spike compression when `neurodamus` is run in Load Balance
Memory mode

## Testing
Due to the nature of the changes, and to standardize them, the test have
been completely reworked with the use of fixture and lots of
improvements to stability Furthermore, the test to distribute cells in
multi-pop and multi cycles has been reworked to use mock object, since
the inherent oscillation in memory measurements could lead to different
results and distributions.

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or
files) are good
* [x] Unit/Scientific test added
* [x] Updated Readme, in-code, developer documentation

---------

Co-authored-by: Fernando Pereira <fernando.pereira@epfl.ch>