Fix Save/Restore coreneuron_input issue: create symlink to the save coreneuron_input folder (#214)
## Context
Users should be able to run several save & restore processes in sequence
for one simulation, as explained in
[wiki](https://github.com/BlueBrain/neurodamus/wiki/Using-CoreNeuron-in-Practice#using-save-restore-for-long-running-simulations).
However, currently the restore uses the direct `coreneuron_input` folder
from its checkpoint's parent directory, so it would lead to error
`coreneuron_input dir not found` for multiple restores such as "save ->
save & restore -> restore" because the 2nd restore only knows its checkpoint folder. 
Therefore, this PR tries to fix it by creating a
symlink pointing to the `coreneuron_input` in the same parent directory as the checkpoint
folder. At the end of the last restore, this link will be removed by
an existing function in `node.py`.

## Scope
Create a `symlink` in `_coreneuron_params` for restore. 

## Testing
Improve `test_save_restore_cli` in `test_cli_opts.py` that it runs 2
restore process in a chain.

## Review
* [x] PR description is complete
* [x] Coding style (imports, function length, New functions, classes or
files) are good
* [x] Unit/Scientific test added
* [ ] Updated Readme, in-code, developer documentation